+++ apps/registration/src/components/ScopeSection.tsx
